Added utils in platform library to provide platform utilities, added ccoSleep for thread sleeping (specifically reducing high CPU load), implemented ccoGetWindowFramebufferSize for win32, CcoViewport now uses f32 for depth ranges, and set C standard to C17

+++ src/graphics/include/graphics/graphics_types.h
@@ -38,8 +38,8 @@ typedef struct {
 typedef struct {
     i32 x, y;
     i32 w, h;
-    u32 minDepth; ///< @brief The smallest depth value allowed (clamps depths smaller than this)
-    u32 maxDepth; ///< @brief The largest depth value allowed (clamps depths larger than this)
+    f32 minDepth; ///< @brief The smallest depth value allowed (clamps depths smaller than this)
+    f32 maxDepth; ///< @brief The largest depth value allowed (clamps depths larger than this)
 } CcoViewport;
 
 /// @brief A definition of a single point in a geometric object
